// fs-frontend-3-2025-71607/src/app/features/panels/patient/models/patientModel.ts
export interface Patient {
  id: string;
  age: number;
  height: number;
  weight: number;
  bloodGroup: string;
  firstName: string;
  lastName: string;
  dateOfBirth: Date;
  nationalIdentity: string;
  phone: string;
  address: string;
  email: string;
  appointmentId: number;
  appointmentDate: Date;
  appointmentTime: string;
  appointmentRapor: string;
}